Aksjehandel del2

+++ Aksjehandel/Controllers/StockController.cs
@@ -1,6 +1,7 @@
 ﻿using aksjehandel.DAL;
 using aksjehandel.Models;
 using Castle.Core.Logging;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
@@ -21,13 +22,19 @@ namespace aksjehandel.Controllers
         private readonly IStockRepository _db;
 
         private ILogger<StockController> _log;
+        private const string _signedIn = "signedIn";
+        private const string _notSignedIn = "";
         public StockController(IStockRepository db, ILogger<StockController> log)
         {
             _db = db;
             _log = log;
         }
-       public async Task<ActionResult> regOrder(Order newOrder)
+        public async Task<ActionResult> regOrder(Order newOrder)
         {
+            if (string.IsNullOrEmpty(HttpContext.Session.GetString(_signedIn)))
+            {
+                return Unauthorized("Ikke logget inn");
+            }
             if (ModelState.IsValid)
             {
                 bool returnOK = await _db.RegOrder(newOrder);
@@ -43,6 +50,10 @@ namespace aksjehandel.Controllers
         }
         public async Task<ActionResult> DeleteOrder(int id)
         {
+            if (string.IsNullOrEmpty(HttpContext.Session.GetString(_signedIn)))
+            {
+                return Unauthorized("Ikke logget inn");
+            }
             bool returnOK = await _db.DeleteOrder(id);
             if (!returnOK)
             {
@@ -53,6 +64,10 @@ namespace aksjehandel.Controllers
         }
         public async Task<ActionResult> ChangeOrder(Order changeOrder)
         {
+            if (string.IsNullOrEmpty(HttpContext.Session.GetString(_signedIn)))
+            {
+                return Unauthorized("Ikke logget inn");
+            }
             if (ModelState.IsValid)
             {
                 bool returnOk = await _db.ChangeOrder(changeOrder);
@@ -68,8 +83,12 @@ namespace aksjehandel.Controllers
         }
         public async Task<ActionResult> GetOneOrder(int id)
         {
+            if (string.IsNullOrEmpty(HttpContext.Session.GetString(_signedIn)))
+            {
+                return Unauthorized("Ikke logget inn");
+            }
             Order oneOrder = await _db.GetOneOrder(id);
-            if(oneOrder == null)
+            if (oneOrder == null)
             {
                 _log.LogInformation("Fant ikke ordren");
                 return NotFound("Fant ikke ordren");
@@ -78,30 +97,74 @@ namespace aksjehandel.Controllers
         }
         public async Task<ActionResult> getAllOrders(int portfolioId)
         {
+            if (string.IsNullOrEmpty(HttpContext.Session.GetString(_signedIn)))
+            {
+                return Unauthorized("Ikke logget inn");
+            }
             List<Order> allOrders = await _db.GetAllOrders(portfolioId);
             return Ok(allOrders);
 
         }
         public async Task<ActionResult> GetAllShareholdings(int portfolioId)
         {
+            if (string.IsNullOrEmpty(HttpContext.Session.GetString(_signedIn)))
+            {
+                return Unauthorized("Ikke logget inn");
+            }
             List<Shareholding> allShareholdings = await _db.GetAllShareholdings(portfolioId);
             return Ok(allShareholdings);
         }
         public async Task<ActionResult> GetAllPortfolios()
         {
+            if (string.IsNullOrEmpty(HttpContext.Session.GetString(_signedIn)))
+            {
+                return Unauthorized("Ikke logget inn");
+            }
             List<Portfolio> allPortfolios = await _db.GetAllPortfolios();
             return Ok(allPortfolios);
         }
         public async Task<ActionResult> GetAllCompanies()
         {
+            if (string.IsNullOrEmpty(HttpContext.Session.GetString(_signedIn)))
+            {
+                return Unauthorized("Ikke logget inn");
+            }
             List<Company> allCompanies = await _db.GetAllCompanies();
             return Ok(allCompanies);
-            
+
         }
         public async Task<ActionResult> GetAllTrades()
         {
+            if (string.IsNullOrEmpty(HttpContext.Session.GetString(_signedIn)))
+            {
+                return Unauthorized("Ikke logget inn");
+            }
             List<Trade> allTrades = await _db.GetAllTrades();
             return Ok(allTrades);
         }
+
+        public async Task<ActionResult> SignIn(User user)
+        {
+            if (ModelState.IsValid)
+            {
+                bool returnOk = await _db.SignIn(user);
+                if (!returnOk)
+                {
+                    _log.LogInformation("Innloggingen feilet for bruker. Brukernavn: " + user.Username);
+                    HttpContext.Session.SetString(_signedIn, _notSignedIn);
+                    return Ok(false);
+                }
+                HttpContext.Session.SetString(_signedIn, _signedIn);
+                return Ok(true);
+            }
+            _log.LogInformation("Feil i inputvalidering");
+            return BadRequest("Feil i inputvalidering på server");
+        }
+
+        public void SignOut()
+        {
+            HttpContext.Session.SetString(_signedIn, _notSignedIn);
+        }
+
     }
 }
